+++ features/bootstrap/FeatureContext.php
@@ -976,12 +976,12 @@ pods:
           SERVER_SCRIPT: '/opt/app/src/server.php'
           from-maps:
             KEY0: 'map1.key0'
-          import-map:
+          import-maps:
             - 'map2'
           from-secrets: #To fetch some value from secret/vault
             KEY1: 'map-vault.key1'
             KEY2: 'map-vault.key2'
-          import-secret:
+          import-secrets:
             - 'map-vault2'
   demo-pods:
     replicas: 1